Consolidate OpenAI Agents message handling
